+++ src/pdf_building.h
@@ -10,6 +10,6 @@ struct pdf {
  * to hold the markdown in pdf form and
  * return a pointer to it.
  */
-int build_pdf_from_md (const char *md,
+size_t build_pdf_from_md (const char *md,
                        size_t md_size,
                        struct pdf *out_pdf);
